# api/openapi.yaml
openapi: 3.0.0
info:
  title: Mindburn Aletheia API
  version: 1.0.0
  description: API for Mindburn Aletheia human-in-the-loop AI verification platform
servers:
  - url: https://api.mindburn.org/{stage}
    variables:
      stage:
        default: dev
        enum:
          - dev
          - staging
          - prod
components:
  securitySchemes:
    CognitoAuth:
      type: apiKey
      name: Authorization
      in: header
      x-amazon-apigateway-authtype: cognito_user_pools
    ApiKeyAuth:
      type: apiKey
      name: x-api-key
      in: header
...
paths:
  /workers:
    get:
      summary: List available workers
      security:
        - CognitoAuth: []
      responses:
        '200':
          description: List of workers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    status:
                      type: string
                    skills:
                      type: array
                      items:
                        type: string
    post:
      summary: Register new worker
      security:
        - CognitoAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - telegram_id
                - skills
              properties:
                telegram_id:
                  type: string
                skills:
                  type: array
                  items:
                    type: string
...